INS-647: cleanup logicrunner tests
